% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{Arima_by_county}
\alias{Arima_by_county}
\title{county-level model}
\usage{
Arima_by_county(
  src,
  state.in = "Massachusetts",
  county.in = "Norfolk",
  MAorder = NULL,
  Difforder = 1,
  basedate = "2020-02-15",
  lookback_days = 29,
  ARorder = NULL,
  max_date = NULL
)
}
\arguments{
\item{src}{a tibble as returned by nytimes_state_data() or jhu_us_data()}

\item{state.in}{character(1) state name}

\item{county.in}{character(1) state name}

\item{MAorder}{numeric(1) order of moving average component, defaults to NULL in which case min_bic is used to
find best value on grid (0:5)^2}

\item{Difforder}{numeric(1) order of differencing d in ARIMA(p,d,q)}

\item{basedate}{character(1) used by lubridate::as_date to filter away all earlier records}

\item{lookback_days}{numeric(1) only uses this many days from most recent in src}

\item{ARorder}{order of autoregressive component, defaults to NULL for optimizing choice, see MAorder}

\item{max_date}{a date from which to start lookback ... defaults to NULL in which
case the latest available date is used}
}
\value{
instance of S3 class Arima_sars2pack
}
\description{
county-level model
}
\examples{
nytc = nytimes_county_data()
Arima_by_county(nytc)
}
