% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt.R
\name{cumulative_events_nyt_county}
\alias{cumulative_events_nyt_county}
\title{create cumulative_events instance for county in state data from NYT}
\usage{
cumulative_events_nyt_county(
  src,
  eventtype = "confirmed",
  statename = "Massachusetts",
  countyname = "Norfolk"
)
}
\arguments{
\item{src}{a tibble}

\item{eventtype}{character(1) 'confirmed' or 'deaths'}

\item{statename}{character(1) state name}

\item{countyname}{character(1) county name}
}
\description{
create cumulative_events instance for county in state data from NYT
}
\examples{
nytc = nytimes_county_data()
norf = cumulative_events_nyt_county(nytc, eventtype = "confirmed",
   statename = "Massachusetts", countyname="Norfolk")
norf
inorf = form_incident_events(trim_from(norf, "2020-03-01"))
plot(inorf)
}
