% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu_appsupp.R
\name{form_incident_events}
\alias{form_incident_events}
\title{transform cumulative data to daily incidence}
\usage{
form_incident_events(cum)
}
\arguments{
\item{cum}{cumulative events instance}
}
\description{
transform cumulative data to daily incidence
}
\examples{
suppressWarnings({dat = enriched_jhu_data()})
cusa = cumulative_events_ejhu(dat, eventtype="confirmed",
 alpha3="USA")
cusa
cusa = trim_from(cusa, "2020-03-01")
iusa = form_incident_events(cusa)
plot(iusa)
}
