% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_06_12}
\alias{min_bic_2020_06_12}
\title{a statewide collection of best AR/MA model orders, 6/12}
\format{
S3 instance
}
\usage{
min_bic_2020_06_12
}
\description{
a statewide collection of best AR/MA model orders, 6/12
}
\keyword{datasets}
