% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_07_06}
\alias{min_bic_2020_07_06}
\title{a statewide collection of best AR/MA model orders, 7/06}
\format{
S3 instance
}
\usage{
min_bic_2020_07_06
}
\description{
a statewide collection of best AR/MA model orders, 7/06
}
\keyword{datasets}
