% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_10_07_death}
\alias{min_bic_2020_10_07_death}
\title{a statewide collection of best AR/MA model orders, 10/07, death only, for debugging}
\format{
S3 instance
}
\usage{
min_bic_2020_10_07_death
}
\description{
a statewide collection of best AR/MA model orders, 10/07, death only, for debugging
}
\keyword{datasets}
