% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_10_21}
\alias{min_bic_2020_10_21}
\title{a statewide collection of best AR/MA model orders, 10/21}
\format{
S3 instance
}
\usage{
min_bic_2020_10_21
}
\description{
a statewide collection of best AR/MA model orders, 10/21
}
\keyword{datasets}
