% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_11_22}
\alias{min_bic_2020_11_22}
\title{a statewide collection of best AR/MA model orders, 11/22}
\format{
S3 instance
}
\usage{
min_bic_2020_11_22
}
\description{
a statewide collection of best AR/MA model orders, 11/22
}
\keyword{datasets}
