% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_11_25}
\alias{min_bic_2020_11_25}
\title{a statewide collection of best AR/MA model orders, 11/25}
\format{
S3 instance
}
\usage{
min_bic_2020_11_25
}
\description{
a statewide collection of best AR/MA model orders, 11/25
}
\keyword{datasets}
