% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_11_26_death}
\alias{min_bic_2020_11_26_death}
\title{a statewide collection of best AR/MA model orders, 11/26, death only}
\format{
S3 instance
}
\usage{
min_bic_2020_11_26_death
}
\description{
a statewide collection of best AR/MA model orders, 11/26, death only
}
\keyword{datasets}
