% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_Rt.R
\name{run_ee}
\alias{run_ee}
\title{basic support, formerly completely hidden in inst, uses Gamma-based MCMC for SI}
\usage{
run_ee(arima_out, ...)
}
\arguments{
\item{arima_out}{instance of Arima_sars2pack}
}
\description{
basic support, formerly completely hidden in inst, uses Gamma-based MCMC for SI
}
