% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dometa.R
\name{run_meta_county}
\alias{run_meta_county}
\title{county within state metaanalysis of incidence velocity}
\usage{
run_meta_county(nytc, state.in = "Massachusetts", verbose = TRUE, ...)
}
\arguments{
\item{nytc}{nytimes_county_data() output}

\item{state.in}{character(1)}

\item{verbose}{logical(1) will print '.' after each county's optimal AR/MA orders found}

\item{\dots}{passed to \code{Arima_by_county}}
}
\description{
county within state metaanalysis of incidence velocity
}
\note{
The search for optimal AR/MA orders may throw warnings.
}
\examples{
nytc = nytimes_county_data()
m1 = run_meta_county(nytc)
names(m1$drifts) = gsub(".drift", "", names(m1$drifts))
rmeta::meta.summaries(m1$drifts, m1$se.drifts)
o = order(m1$drifts)
rmeta::metaplot(m1$drifts[o], m1$se.drifts[o], labels=names(m1$drifts)[o], cex=.7, 
  xlab="Infection velocity (CHANGE in number of confirmed cases/day)", ylab="County")
}
