% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_cens_G}
\alias{si_cens_G}
\title{an EpiEstim-based, MCMC-derived SI sample, with Gamma basis}
\format{
list
}
\usage{
si_cens_G
}
\description{
an EpiEstim-based, MCMC-derived SI sample, with Gamma basis
}
\keyword{datasets}
